# web-backend/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Simanfor
  version: '2.0.2'
servers:
  - url: 'http://35.246.161.83/simanfor/'
    description: Dev
  - url: 'http://localhost:3000/'
    description: Local
paths:
...
  '/inventory/{inventoryId}':
    parameters:
      - schema:
          type: string
        name: inventoryId
        in: path
        required: true
    get:
      summary: Get an inventory
      tags: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Inventory'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    default: false
                  message:
                    type: string
                  code:
                    default: 404
                  string_code:
                    type: string                    
      operationId: get-inventory-inventoryId
      description: Gets an inventory.
    delete:
      summary: 'Delete an inventory'
      operationId: delete-inventory-inventoryId
      responses:
        '204':
          description: No content
        '404':
          description: Not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    default: false
                  message:
                    type: string
                  code:
                    default: 404
                  string_code:
                    type: string           
      description: Removes an inventory.
...
components:
  schemas:
...
    Inventory:
      title: Inventory
      type: object
      properties:
        _id:
          type: string
        inventoryId:
          type: string
        name:
          type: string
        year:
          type: string
        creator:
          type: string
        creatorId:
          type: string
        creationDate:
          type: number
        public:
          type: boolean
        type:
          type: string
        fileUrl:
          type: string
...
  securitySchemes:
    Authorization:
      name: Authorization
      type: apiKey
      in: header
security:
  - Authorization: []